refactor: rvr 64bit address ptr (#2928)
refactor address pointer related variables to 64bit instead of 32bit.
this PR modifies multiple files:
1) rvr core c and trace headers
2) ffi files
3) rvr code for extensions


reth benchmark result (original 32bit address space variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28121173062
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122726878

reth benchmark result (64bit address pointer variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28119229848
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122157089

By comparing "Exec Median (ms)" columns, it seems that 64bit version is
`10ms` faster or slower compared with the original 32bit version (in
both pure and metered modes). I am not sure why, it could be system
noises affecting it. Also, the comparison only compares execution speed,
excluding other possible important metrics when evaluating this pr.

resolves INT-8321

---------

+++ crates/rvr/rvr-openvm/c/openvm_check_mem_bounds.h
@@ -6,13 +6,13 @@
 #include "openvm_constants.h"
 #include "openvm_util.h"
 
-__attribute__((noreturn, cold)) void abort_oob(uint32_t start, size_t size,
+__attribute__((noreturn, cold)) void abort_oob(uint64_t start, size_t size,
                                                size_t mem_size);
 
 static constexpr size_t OPENVM_MEM_SIZE = (size_t)MEMORY_MASK + 1u;
 
 static __attribute__((always_inline)) inline void check_mem_bounds_u8(
-    uint32_t start) {
+    uint64_t start) {
   if (unlikely(start > OPENVM_MEM_SIZE - sizeof(uint8_t))) {
     abort_oob(start, sizeof(uint8_t), OPENVM_MEM_SIZE);
   }
@@ -20,7 +20,7 @@ static __attribute__((always_inline)) inline void check_mem_bounds_u8(
 }
 
 static __attribute__((always_inline)) inline void check_mem_bounds_i8(
-    uint32_t start) {
+    uint64_t start) {
   if (unlikely(start > OPENVM_MEM_SIZE - sizeof(int8_t))) {
     abort_oob(start, sizeof(int8_t), OPENVM_MEM_SIZE);
   }
@@ -28,7 +28,7 @@ static __attribute__((always_inline)) inline void check_mem_bounds_i8(
 }
 
 static __attribute__((always_inline)) inline void check_mem_bounds_u16(
-    uint32_t start) {
+    uint64_t start) {
   if (unlikely(start > OPENVM_MEM_SIZE - sizeof(uint16_t))) {
     abort_oob(start, sizeof(uint16_t), OPENVM_MEM_SIZE);
   }
@@ -36,7 +36,7 @@ static __attribute__((always_inline)) inline void check_mem_bounds_u16(
 }
 
 static __attribute__((always_inline)) inline void check_mem_bounds_i16(
-    uint32_t start) {
+    uint64_t start) {
   if (unlikely(start > OPENVM_MEM_SIZE - sizeof(int16_t))) {
     abort_oob(start, sizeof(int16_t), OPENVM_MEM_SIZE);
   }
@@ -44,7 +44,7 @@ static __attribute__((always_inline)) inline void check_mem_bounds_i16(
 }
 
 static __attribute__((always_inline)) inline void check_mem_bounds_u32(
-    uint32_t start) {
+    uint64_t start) {
   if (unlikely(start > OPENVM_MEM_SIZE - sizeof(uint32_t))) {
     abort_oob(start, sizeof(uint32_t), OPENVM_MEM_SIZE);
   }
@@ -52,7 +52,7 @@ static __attribute__((always_inline)) inline void check_mem_bounds_u32(
 }
 
 static __attribute__((always_inline)) inline void check_mem_bounds_u64(
-    uint32_t start) {
+    uint64_t start) {
   if (unlikely(start > OPENVM_MEM_SIZE - sizeof(uint64_t))) {
     abort_oob(start, sizeof(uint64_t), OPENVM_MEM_SIZE);
   }
@@ -60,14 +60,14 @@ static __attribute__((always_inline)) inline void check_mem_bounds_u64(
 }
 
 static __attribute__((always_inline)) inline void check_mem_bounds_range(
-    uint32_t start, size_t size) {
+    uint64_t start, size_t size) {
   if (unlikely(start > OPENVM_MEM_SIZE || size > OPENVM_MEM_SIZE - start)) {
     abort_oob(start, size, OPENVM_MEM_SIZE);
   }
 }
 
 static __attribute__((always_inline)) inline void check_mem_bounds_u64_range(
-    uint32_t base_addr, uint32_t num_words) {
+    uint64_t base_addr, uint32_t num_words) {
   check_mem_bounds_range(base_addr, (size_t)num_words * sizeof(uint64_t));
 }
